Moved _GenerateDeviceID() from Configuration to Inventory
